refactor: split out side bar functions to separate classes
This will scale better when we add yet more functions.
